move entity declarations to their own files
