+++ backend/src/app.module.ts
@@ -1,18 +1,16 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppService } from './app.service';
-import {
-  User,
-  Event,
-  EventManager,
-  InvitedUsers,
-  EventLocation,
-  EventOccurrence,
-  Registration,
-  Ticket,
-  Notification,
-  Invitation,
-} from './entity/entities';
+import { Invitation } from './entity/Invitation';
+import { Notification } from './entity/Notification';
+import { Ticket } from './entity/Ticket';
+import { Registration } from './entity/Registration';
+import { EventOccurrence } from './entity/EventOccurrence';
+import { EventLocation } from './entity/EventLocation';
+import { InvitedUsers } from './entity/InvitedUsers';
+import { EventManager } from './entity/EventManager';
+import { Event } from './entity/Event';
+import { User } from './entity/User';
 import { UsersModule } from './users/users.module';
 import { AuthModule } from './auth/auth.module';
 import * as path from 'path';
