+++ backend/src/auth/jwtData.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { JwtPayload } from './entity/jwtPayload.entity';
+import { JwtPayload } from './entity/JwtPayload';
 import { Request } from 'express';
 
 interface RequestWithUser extends Request {
