+++ backend/src/entity/User.ts
@@ -1,0 +1,20 @@
+import { Entity, Column, OneToMany } from 'typeorm';
+import { BaseEntity } from './BaseEntity';
+import { RefreshToken } from './RefreshToken';
+
+@Entity()
+export class User extends BaseEntity {
+  @Column()
+  fullName: string;
+
+  @Column({ unique: true })
+  email: string;
+
+  @Column()
+  passwordHash: string;
+
+  @OneToMany(() => RefreshToken, (refreshToken) => refreshToken.user, {
+    cascade: true,
+  })
+  refreshTokens: RefreshToken[];
+}
